+++ examples/scatter/BubbleScatterChart.tsx
@@ -11,13 +11,14 @@ import {
     Tooltip,
     TooltipData,
     TooltipDataItem,
+    ViewController,
 } from '@chsk/core'
 import { Scatter, ScatterPoints, ScatterCrosshair, isScatterData } from '@chsk/xy'
-import { downloadTheme } from '@chsk/themes'
+import { buttonTheme } from '@chsk/themes'
 import { generateXYValues } from './generators'
 import { generateMixedPopulation, randomNormalValue, round1dp, round3dp } from '../utils'
 import { MilestoneStory } from '../types'
-import { DownloadButtons } from '../navigation'
+import { DownloadButtons, IconButton } from '../navigation'
 
 export const generateBubbleScatterData = () => {
     const x1 = generateMixedPopulation([80, 10, 10], [1, 3, 5], [1, 1, 1])
@@ -48,7 +49,7 @@ export const generateBubbleScatterData = () => {
     ]
 }
 
-const customTheme: ThemeSpec = mergeTheme(downloadTheme, {
+const customTheme: ThemeSpec = mergeTheme(buttonTheme, {
     line: {
         grid: {
             stroke: '#dddddd',
@@ -175,6 +176,7 @@ export const BubbleScatterChart = ({ fref, chartData, rawData }: MilestoneStory)
                 <Typography position={[-45, -30]} variant={'title'}>
                     Large scatter plot
                 </Typography>
+                <ViewController component={IconButton} />
                 <DownloadButtons position={[630, -30]} data image />
             </Scatter>
         </Chart>